Allow deactivating atomicity of migrations
This is very useful when you want to use the 'CONCURRENTLY' option of PostgreSQL (see http://www.postgresql.org/docs/9.1/static/sql-createindex.html#SQL-CREATEINDEX-CONCURRENTLY), which requires the query to be run out of a transaction block. We've run on that problem several times in our team and couldn't find a way to handle our migrations within South.

With that PR, the way to do it would be to set `atomic = False` in the transaction class and then to define a RunSQL operation with the query and the correct state_operations.

Tests are coming.